requirements.txt
python-dotenv
requests
Pillow
pycryptodome
matplotlib